Separating I/O from Game class.
