+++ README.md
@@ -3,3 +3,7 @@
 Hunt the Wumpus is a classic game [originally developed by Gregory Yob](http://www.atariarchives.org/bcc1/showpage.php?page=247).
 
 The objective of this project is to develop a text-based, Ruby version of the game as a group exerise as part of [Launch School](https://launchschool.com)
+
+### How to run the program:
+
+Simply run `ruby wumpus_game.rb` from the root directory of the project while in your terminal/command prompt.
