--- game.rb
@@ -1,4 +1,0 @@
-require_relative 'app/wumpus'
-
-game = WumpusGame.new
-game.start
