+++ messages.yaml
@@ -1,0 +1,7 @@
+welcome: Welcome to Hunt the Wumpus!
+goodbye: Thanks for Playing Hunt the Wumpus. Goodbye!
+move_choice: "What would you like to do? (s)hoot or (m)ove:"
+pick_a_room: "Pick a room to move to: "
+shot_distance: Choose a shot distance in number of rooms from 1-5
+room_target: "Draw your arrow path by selecting a room: "
+current_room: "You are currently in room "
